More type tightening for game and server.

+++ src/game/default/g_items.h
@@ -28,12 +28,12 @@ extern g_item_t g_items[];
 
 #define ITEM_INDEX(x) ((x) - g_items)
 
-boolean_t G_AddAmmo(g_edict_t *ent, g_item_t *item, int count);
+boolean_t G_AddAmmo(g_edict_t *ent, g_item_t *item, short count);
 g_edict_t *G_DropItem(g_edict_t *ent, g_item_t *item);
 g_item_t *G_FindItem(const char *pickup_name);
 g_item_t *G_FindItemByClassname(const char *class_name);
 void G_InitItems(void);
-g_item_t *G_ItemByIndex(int index);
+g_item_t *G_ItemByIndex(unsigned short index);
 void G_PrecacheItem(g_item_t *it);
 void G_ResetFlag(g_edict_t *ent);
 void G_SetItemNames(void);
